refactoring wip 1: multiple modules
