+++ src/logging.ts
@@ -1,0 +1,7 @@
+/*
+Logging facilities for the extension.
+*/
+
+export function log (level: string, message: string) {
+	console.log('KataPod ' + level.toUpperCase() + ' ' + message);
+}
