+++ src/runCommands.ts
@@ -1,0 +1,11 @@
+/*
+Tools to run commands and scripts on the terminals.
+*/
+
+// import * as vscode from 'vscode';
+// const path = require('path');
+
+// import {log} from './logging';
+// import {getWorkingDir} from './filesystem';
+
+
